+++ saleor/graphql/schema.graphql
@@ -24342,6 +24342,18 @@ input PermissionGroupCreateInput {
   """List of users to assign to this group."""
   addUsers: [ID!]
 
+  """
+  Determine if group have restricted access to channels.
+  
+  Added in Saleor 3.11.
+  
+  Note: this API is currently in Feature Preview and can be subject to changes at later point.
+  """
+  restrictedAccessToChannel: Boolean!
+
+  """List of channels to assign as accessible for this group."""
+  addChannels: [ID!]
+
   """Group name."""
   name: String!
 }
@@ -24364,6 +24376,18 @@ input PermissionGroupUpdateInput {
   """List of users to assign to this group."""
   addUsers: [ID!]
 
+  """
+  Determine if group have restricted access to channels.
+  
+  Added in Saleor 3.11.
+  
+  Note: this API is currently in Feature Preview and can be subject to changes at later point.
+  """
+  restrictedAccessToChannel: Boolean!
+
+  """List of channels to assign as accessible for this group."""
+  addChannels: [ID!]
+
   """Group name."""
   name: String
 
@@ -24372,6 +24396,9 @@ input PermissionGroupUpdateInput {
 
   """List of users to unassign from this group."""
   removeUsers: [ID!]
+
+  """List of channels to unassign as accessible for this group."""
+  removeChannels: [ID!]
 }
 
 """
